I've got one extra thing, but I have only tested this on windows so far. I'm not able to test it on MacOS, Linux, I can do.
This change ensures that when the last window closes, the (DLL) library is cleaned up and can also be re-initialized.
Why is this important?

Some environments (like the DrRacket IDE) keep DLLs loaded (even when one requests to unload them).

After closing the last window, the DLL stays in 'exit' state and it is not possible anymore to create new windows.

This change makes sure that the shared library keeps working while loaded.