Fix overwritten event entries in _webui_new_event_inf
When creating new event info data and number, search for an empty event array entry rather than simply increasing the number and potentially overwriting existing info. This prevents intermittent double deletions when event blocking is disabled and many events of varying run durations occur in a short time.

Also, increase the number of event entries that can be used to WEBUI_MAX_IDS which is the allocated events variable size.