Fix bootstrap problem with logger function.
To intercept this with a log handler, it must be possible
to set the log handler before any logging takes place.

This logger_func / logger_user_data cannot be part of webui_core_t,
because logging is already used in _webui_init() and _webui of
type webui_core_t gets reset to 0 in _webui_init().


Signed-off-by: Hans Dijkema <hans@dijkewijk.nl>